further physics isolation
